Implement ⍕ (format) and ⍸ (where); fix character array display
Add two primitives via the six-site wiring checklist:

⍕ format — monadic renders any value as its character representation,
reusing boxdisplay so it inherits existing rendering; multi-line results
become a character matrix padded to the widest row. Dyadic A⍕B takes
scalar decimals or a width/decimals pair, renders negatives with the APL
high minus, and signals DOMAIN ERROR when the field is too narrow.

⍸ where — monadic yields the positions of 1s, ⎕IO-shifted in the parser
alongside ⍳/⍋/⍒ since the result is indices. Rank ≥ 2 returns a nested
vector of per-axis subscript vectors; non-boolean input is DOMAIN ERROR.

Also fixes a pre-existing display bug that ⍕ made obvious: simple
character arrays were printed space-separated ('hello' as h e l l o).
GNU APL separates only numeric arrays, so boxdisplay now selects the
separator per element type. The REPL kept a second divergent copy of the
display logic in main.rs and only routed nested/matrix values through
boxdisplay, so character vectors bypassed the fix — main.rs now routes
them to the shared renderer too.

346 tests green, zero clippy warnings (also clears the long-standing
needless_borrow in encode_decode.rs), fmt clean.

+++ src/squad.rs
@@ -2,7 +2,6 @@
 //!
 //! Mirrors `Bif_F2_INDEX` and `Bif_F12_ROTATE1` in C++ (simplified).
 
-use crate::cell::Cell;
 use crate::shape::Shape;
 use crate::types::{AplResult, ErrorCode};
 use crate::value::ValueP;
@@ -28,7 +27,7 @@ pub fn squad(a: &ValueP, b: &ValueP) -> AplResult<ValueP> {
         .collect::<Result<Vec<_>, _>>()?;
     // Bounds check
     for (i, &idx) in indices.iter().enumerate() {
-        let axis_len = if i < rank as usize {
+        let axis_len = if i < rank {
             b.get_shape_item(i as i16)
         } else {
             1
@@ -110,6 +109,7 @@ pub fn rotate_first(a: &ValueP, b: &ValueP) -> AplResult<ValueP> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::cell::Cell;
 
     #[test]
     fn test_squad_vector() {
